+++ contao/languages/de/default.php
@@ -25,6 +25,7 @@ $GLOBALS['TL_LANG']['CTE']['ct_sliderElement'][0] = 'Slider-Element';
 $GLOBALS['TL_LANG']['CTE']['ct_teaserBox'][0] = 'Teaser Box';
 $GLOBALS['TL_LANG']['CTE']['ct_wrapperStart'][0] = 'Inhaltscontainer Anfang';
 $GLOBALS['TL_LANG']['CTE']['ct_wrapperStop'][0] = 'Inhaltscontainer Ende';
+$GLOBALS['TL_LANG']['CTE']['feature_section'][0] = 'Highlight-Bereich';
 
 $GLOBALS['TL_LANG']['WCAG']['enable_high_contrast'] = ['Kontrast umschalten', 'Hier können Sie den Modus mit hohem Kontrast umschalten.'];
 $GLOBALS['TL_LANG']['WCAG']['enable_font_size'] = ['Schriftgröße umschalten', 'Hier gelangen Sie zu einer Seite mit Erklärungen, wie Sie in modernen Browsern die Schriftgröße anpassen können.'];