+++ contao/languages/en/tl_page.php
@@ -17,9 +17,14 @@ declare(strict_types=1);
  */
 
 $GLOBALS['TL_LANG']['tl_page']['wcag_legend'] = 'WCAG-Settings (Web Content Accessibility Guidelines)';
+$GLOBALS['TL_LANG']['tl_page']['theme_components_legend'] = 'Theme component settings';
 
 $GLOBALS['TL_LANG']['tl_page']['enable_high_contrast'] = ['Allow switching to high contrast mode', 'Allows switching to the high contrast mode. This option is enabled by default.'];
 $GLOBALS['TL_LANG']['tl_page']['enable_auto_high_contrast'] = ['Automatic switch to high-contrast mode', 'Automatically switches to high-contrast mode when the system settings change, for example when Dark Mode is enabled or disabled.'];
 $GLOBALS['TL_LANG']['tl_page']['enable_font_size'] = ['Allow display of the icon for changing the font size', 'Allows to display the icon for changing the font size. This function is implemented here by linking to a page with information about changing the font size. The font size itself is not changed.'];
 $GLOBALS['TL_LANG']['tl_page']['enable_font_size_switcher'] = ['Allow display of font size toggle (-A A A+) for changing font size', 'Allows showing the font size toggle for changing the font size. This function directly changes the font size.'];
 $GLOBALS['TL_LANG']['tl_page']['font_size_comment_page'] = ['Select here the page on which you carry out your explanations of the font size', 'Since all modern browsers support resizing text on their own, this feature should no longer be implemented through Javascript or stylesheets. Instead, it is easier to provide information about changing the font size on a separate page.'];
+$GLOBALS['TL_LANG']['tl_page']['animateOnScroll'] = [
+    'Enable scroll animation effects (slide-in)',
+    'If enabled, the required CSS and JavaScript files will be loaded. Elements can use the classes reveal, reveal-left, reveal-right or reveal-up to create a slide-in effect when scrolling.'
+];