+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Types of changes
     Security in case of vulnerabilities.
 )
 
+## [1.9.0](https://github.com/contao-themes-net/theme-components-bundle/tree/1.9.0) – 2026-07-22
+
+- [Added] New content element "Feature Section"
+- [Added] New content element "Icon Tile" (based on the hyperlink element with image and customized template)
+
 ## [1.8.3](https://github.com/contao-themes-net/theme-components-bundle/tree/1.8.3) – 2026-01-05
 
 - [Fixed] Remove warning `Attempt to read property "rootId" on null`